auth_platform/auth_platform/auth_service/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
pyjwt
pytest
httpx
pyotp
